requirements.txt
opencv-contrib-python==4.0.0.21
imutils==0.5.2
numpy==1.16.2
scipy==1.2.1
xmltodict==0.12.0
PyYAML==5.1
